v0.3.3 Fixed issue in FF disconnecting after a few seconds
FF seems to GC data channels which should not be GC'd. Added
a hard reference to further signal that these channels are
still in use. Will investigate further to see if this was an
implementation error on my part, or if FF actually has a bug
which needs to be addressed.